\name{list2hexList}
\alias{list2hexList}
\title{Convert list to hexList}
\description{
  Converts a list of hexbin objects with same xbnds, ybnds, shape and
  xbins to a \code{\link{hexList}} object.
}
\usage{
list2hexList(binlst)
}
\arguments{
  \item{binlst}{A list of hexbin objects}
}

\value{
 a \code{\link{hexList}} object
}

\author{Nicholas Lewin-Koh}


\seealso{\code{\link{hexList}},\code{\link{hdiffplot}} }

\keyword{misc}

